Should improve performance of MoE models with CUDA significantly. Also improved the rearrangement of the rows in the CUDA backend with custom kernels instead of memcpys, that's about 50% of the speedup here.

| GPU         | Model           | Test   |   t/s master |   t/s sl/moe-rework-2 |   Speedup |
|:------------|:----------------|:-------|-------------:|----------------------:|----------:|
| RTX 3090 Ti | mixtral Q3_K_S | pp512  |       387.58 |               1226.11 |      3.16 |
| RTX 3090 Ti | mixtral Q3_K_S | tg128  |        43.07 |                 50.40 |      1.17 |