@Dampfinchen 2024-04-06
Alright, wow. PP went down from 11,85 ms/t to 4,95 ms/t. Simply incredible, but I'm not surprised anymore as Slaren always delivers. Llama.cpp's MOE implementation is now extremly robust.