@JohannesGaessler 2024-04-07
Some quick performance comparisons from me:

| GPU        | Model               | Test     |     t/s master |     t/s sl/moe-rework-2 |     Speedup |
| :--------- | :----------------   | :------- | -------------: | ----------------------: | ----------: |
| RTX 3090   | Mixtral 8x7b Q3_K_S | pp512    |         399.25 |                 1173.12 |        2.94 |
| RTX 3090   | Mixtral 8x7b Q3_K_S | tg128    |          53.47 |                   55.80 |        1.04 |
| P40        | Mixtral 8x7b Q3_K_S | pp512    |         187.27 |                  250.23 |        1.34 |
| P40        | Mixtral 8x7b Q3_K_S | tg128    |          23.50 |                   23.91 |        1.02 |

>I had already tested most of these, but on my 3090 I didn't see a meaningful improvement. Anyway I have pushed my changes that I think already cover all of that.

I am measuring a performance difference from the changes:

| GPU        | Model               | Test     |     t/s ea2b7953 |     t/s sl/moe-rework-2 |     Speedup |
| :--------- | :----------------   | :------- | ---------------: | ----------------------: | ----------: |
| RTX 3090   | Mixtral 8x7b Q3_K_S | pp512    |          1112.34 |                 1173.12 |        1.05 |
| RTX 3090   | Mixtral 8x7b Q3_K_S | tg128    |            55.45 |                   55.80 |        1.01 |
| P40        | Mixtral 8x7b Q3_K_S | pp512    |           246.34 |                  250.23 |        1.02 |
| P40        | Mixtral 8x7b Q3_K_S | tg128    |            23.82 |                   23.91 |        1.00 |

>In the long term the goal is to use a grouped GEMM with cutlass without requiring a synchronization. I think that this will also allow removing the row rearrangement entirely, which has a significant cost.

That would definitely help. When you look into this I think it would also make sense to check whether it is possible to set the input/output type to FP32 to avoid the conversion of some tensors. (With the input I suspect that it's probably not possible though.)