@askmyteapot 2024-04-07
Definite speedup on P40 with larger iq4_xs quant with partial offloading. 
```
ggml_cuda_init: GGML_CUDA_FORCE_MMQ:   yes
ggml_cuda_init: CUDA_USE_TENSOR_CORES: no
ggml_cuda_init: found 1 CUDA devices:
  Device 0: Tesla P40, compute capability 6.1, VMM: yes
| model                          |       size |     params | backend    | ngl | test       | updatedPR    t/s | moe-rework-2 t/s | speedup |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ---------- | ---------------: | ---------------: | ------: |
| llama 7B IQ4_XS - 4.25 bpw     |  23.42 GiB |    46.70 B | CUDA       |  28 | pp 4096    |    115.45 ± 0.08 |    106.83 ± 0.07 |  1.080x |
| llama 7B IQ4_XS - 4.25 bpw     |  23.42 GiB |    46.70 B | CUDA       |  28 | tg 128     |     14.52 ± 0.01 |     13.34 ± 0.01 |  1.088x |
```