@slaren 2024-04-08
@ggerganov I really do not want to have to modify 21 functions in exactly the same way again, I would rather spend some time refactoring. Did you find any reason that would prevent making the Metal `kernel_mul_mv_id` kernels a template?