@ggerganov 2024-04-09
> Did you find any reason that would prevent making the Metal kernel_mul_mv_id kernels a template?

No reason at all - I simply wasn't able to fit this into templates. Thanks for doing it - it was very ugly before