@slaren 2024-04-16
It's still missing a Metal implementation. It should be good for CPU and CUDA already.