@slaren 2024-04-18
@ggerganov Do you know why the ggml-ci cuda-v100 failed? The log ends during a `quantize`. Was it a timeout?