@ggerganov 2024-04-18
Yes, it exceeded 30 min. On `master` we were at ~27 min

We can either increase to 40 min or maybe not run `ctest` in Debug?