@NeoZhangJianyu 2024-04-18
> @NeoZhangJianyu @airMeng This change will break mul_mat_id in SYCL again. Sorry for the inconvenience, the change to the interface was necessary to improve performance.

Got it! I will study and fix later.
I have thought to have a rest after fix mul_mat_id() UT in last weekend. :)

Thank for reminding!